central-server/requirements.txt
# Core
fastapi
uvicorn
python-multipart
pydantic_settings

# WebSocket & Real-Time
websockets
aioredis # Optional for Redis pub/sub

# MQTT Support
paho-mqtt

# WebRTC Support
aiortc
aioice

# Video Processing (if doing any server-side decoding)
numpy
opencv-python-headless  # Only if processing frames

# Async
anyio
httpx

# Utilities
python-dotenv
loguru  # Better logging
orjson  # Faster JSON

# Production Extras
gunicorn  # For production deployment
uvloop  # Faster event loop (Linux only)